pyproject.toml
[tool.poetry]
name = "test-task-apptrix"
version = "0.1.0"
description = ""
authors = ["Дима Бамбагаев <bambagaevdmitrij@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_apptrix"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-restframework = "^0.0.1"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
pillow = "^10.0.0"
drf-base64 = "^2.0"
djangorestframework-simplejwt = "^5.2.2"
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
